+++ TechMed.BL/Repository/Interfaces/IPatientCaseRepository.cs
@@ -20,7 +20,7 @@ namespace TechMed.BL.Repository.Interfaces
         Task<PatientCaseWithDoctorVM> GetPatientQueueDetails(int PHCID, int PatientID);
         Task<PatientReferToDoctorVM> PostPatientReferToDoctor(PatientReferToDoctorVM patientReferToDoctorVM);
         Task<PatientFeedbackDTO> PostPatientFeedBack(PatientFeedbackDTO patientFeedback);
-        Task<PatientCaseVM> GetPatientCaseDetailsByCaseID(int PatientCaseID, string contentRootPath);
+        Task<PatientCaseVM> GetPatientCaseDetailsByCaseID(long PatientCaseID, string contentRootPath);
         Task<PatientCaseLevelDTO> GetPatientCaseLevels(int patientID);
         Task<List<PatientCaseDocDTO>> GetPatientCaseDocList(int PatientCaseID, string rootUrl);
         bool IsPatientCaseExist(PatientCaseCreateVM patientCase);
